+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix: `config defaults` should be able to set defaults for any parameter, even if it is for a backend not currently in use.
 * Add `aerolab upgrade [--edge]` command to allow upgrading to latest stable (or latest pre-release).
 * Windows: catch the use of command prompt and warn against it.
+* Improvement: replace the `mesh/multicast` fix system with `aeroconf` parser.
 * TODO: Monitor agi instance states and react accordingly - sizing or cycling from spot to on-demand types.
 * TODO: Document agi instance state monitor.